requirements.txt
contourpy
cuda-pathfinder==1.5.2
cupy-cuda13x==14.0.1
cycler==0.12.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.4.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx
numpy
packaging==26.0
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
typing_extensions==4.15.0
